+++ src/Controller/BO/SocieteController.php
@@ -206,7 +206,10 @@ class SocieteController extends AbstractController
         Invitator $invitator,
         EntityManagerInterface $em
     ) {
-        if (!$this->isCsrfTokenValid('send-invitation-admin', $request->get('token'))) {
+        if (
+            !$this->isCsrfTokenValid('send-invitation-admin', $request->get('token')) ||
+            $societeUser->getSociete() !== $societe
+        ) {
             throw new BadRequestHttpException('Csrf token invalid');
         }
 
@@ -217,6 +220,48 @@ class SocieteController extends AbstractController
             'Un email avec un lien d\'invitation a été envoyé à l\'administrateur "%s".',
             $societeUser->getInvitationEmail()
         ));
+
+        return $this->redirectToRoute('app_bo_societe', [
+            'id' => $societe->getId(),
+        ]);
+    }
+
+    /**
+     * @Route(
+     *      "/societes/{societeId}/invitation/{societeUserId}/supprimer",
+     *      name="app_bo_societe_invite_delete",
+     *      methods={"POST"},
+     *      requirements={"id"="\d+"}
+     * )
+     *
+     * @ParamConverter("societe", options={"id" = "societeId"})
+     * @ParamConverter("societeUser", options={"id" = "societeUserId"})
+     */
+    public function societeInviteDelete(
+        Request $request,
+        Societe $societe,
+        SocieteUser $societeUser,
+        EntityManagerInterface $em
+    ) {
+        if (
+            !$this->isCsrfTokenValid('delete-invitation-admin', $request->get('token')) ||
+            $societeUser->getSociete() !== $societe
+        ) {
+            throw new BadRequestHttpException('Csrf token invalid');
+        }
+
+        if (!$societeUser->getInvitationToken()){
+            $this->addFlash('warning', 'Vous ne pouvez pas supprimer une invitation acceptée.');
+        } else {
+            $invitationEmail = $societeUser->getInvitationEmail();
+            $em->remove($societeUser);
+            $em->flush();
+
+            $this->addFlash('success', sprintf(
+                'L\'invitation de "%s" a été supprimée avec succès.',
+                $invitationEmail
+            ));
+        }
 
         return $this->redirectToRoute('app_bo_societe', [
             'id' => $societe->getId(),